SolutionAnalyzers - Updated net analyers setting AnalysisMode to AllEnabledByDefault and resolved new analyzer builder errors

+++ src/Rhinobyte.Extensions.Reflection/IntermediateLanguage/SignedByteInstruction.cs
@@ -2,6 +2,10 @@
 
 namespace Rhinobyte.Extensions.Reflection.IntermediateLanguage
 {
+	/// <summary>
+	/// An instruction with a corresponding signed bye value
+	/// </summary>
+	/// <remarks>The <see cref="SignedByteInstruction.Value"/> has property type <see cref="short" /> to ensure common language specification compliance</remarks>
 	public sealed class SignedByteInstruction : InstructionBase
 	{
 		internal SignedByteInstruction(int index, int offset, OpCode opcode, sbyte value)
@@ -13,7 +17,8 @@ namespace Rhinobyte.Extensions.Reflection.IntermediateLanguage
 		/// <summary>
 		/// The <see cref="sbyte"/> value of the instruction.
 		/// </summary>
-		public sbyte Value { get; }
+		/// <remarks>Property type is <see cref="short"/> to ensure common language specification compliance</remarks>
+		public short Value { get; }
 
 		public override string ToString()
 			=> $"{base.ToString()}  [SByte Value: {Value}]";
